+++ src/test/java/com/github/ddth/kafka/qnd/QndProducerExample.java
@@ -1,5 +1,8 @@
 package com.github.ddth.kafka.qnd;
 
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+
 import com.github.ddth.kafka.KafkaClient;
 import com.github.ddth.kafka.KafkaMessage;
 
@@ -7,17 +10,19 @@ public class QndProducerExample {
 
     public static void main(String[] args) throws Exception {
         final String zkConnString = "localhost:2181/kafka";
-        final String topic = "testtopic";
+        final String topic = "demo";
 
         KafkaClient kafkaClient = new KafkaClient(zkConnString);
         kafkaClient.init();
 
-        for (int i = 0; i < 100; i++) {
-            byte[] content = String.valueOf(System.currentTimeMillis()).getBytes();
-            kafkaClient.sendMessage(new KafkaMessage().topic(topic).content(content));
+        for (int i = 0; i < 10; i++) {
+            String msg = i + ":" + System.currentTimeMillis();
+            byte[] content = msg.getBytes();
+            Future<KafkaMessage> result = kafkaClient.sendMessage(new KafkaMessage().topic(topic)
+                    .content(content));
+            System.out.println(result.get(1000, TimeUnit.MILLISECONDS));
         }
 
         kafkaClient.destroy();
     }
-
 }